Split notifications into a layout and index route
TanStack Router nests unsubscribe under /notifications, so the subscribe
page cannot own that path. A layout with an Outlet keeps unsubscribe from
rendering inside the subscribe form.
