+++ src/routeTree.gen.ts
@@ -12,6 +12,7 @@ import { Route as rootRouteImport } from './routes/__root'
 import { Route as IndexRouteImport } from './routes/index'
 import { Route as DelegationRouteImport } from './routes/delegation'
 import { Route as NotificationsRouteImport } from './routes/notifications'
+import { Route as NotificationsIndexRouteImport } from './routes/notifications.index'
 import { Route as NotificationsUnsubscribeRouteImport } from './routes/notifications.unsubscribe'
 import { Route as ProposalProposalIdRouteImport } from './routes/proposal.$proposalId'
 import { Route as ApiCronNotificationsRouteImport } from './routes/api/cron.notifications'
@@ -32,11 +33,17 @@ const NotificationsRoute = NotificationsRouteImport.update({
   path: '/notifications',
   getParentRoute: () => rootRouteImport,
 } as any)
-const NotificationsUnsubscribeRoute = NotificationsUnsubscribeRouteImport.update({
-  id: '/notifications/unsubscribe',
-  path: '/notifications/unsubscribe',
-  getParentRoute: () => rootRouteImport,
+const NotificationsIndexRoute = NotificationsIndexRouteImport.update({
+  id: '/',
+  path: '/',
+  getParentRoute: () => NotificationsRoute,
 } as any)
+const NotificationsUnsubscribeRoute =
+  NotificationsUnsubscribeRouteImport.update({
+    id: '/unsubscribe',
+    path: '/unsubscribe',
+    getParentRoute: () => NotificationsRoute,
+  } as any)
 const ProposalProposalIdRoute = ProposalProposalIdRouteImport.update({
   id: '/proposal/$proposalId',
   path: '/proposal/$proposalId',
@@ -47,27 +54,29 @@ const ApiCronNotificationsRoute = ApiCronNotificationsRouteImport.update({
   path: '/api/cron/notifications',
   getParentRoute: () => rootRouteImport,
 } as any)
-const ApiNotificationsTelegramRoute = ApiNotificationsTelegramRouteImport.update({
-  id: '/api/notifications/telegram',
-  path: '/api/notifications/telegram',
-  getParentRoute: () => rootRouteImport,
-} as any)
+const ApiNotificationsTelegramRoute =
+  ApiNotificationsTelegramRouteImport.update({
+    id: '/api/notifications/telegram',
+    path: '/api/notifications/telegram',
+    getParentRoute: () => rootRouteImport,
+  } as any)
 
 export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
   '/delegation': typeof DelegationRoute
-  '/notifications': typeof NotificationsRoute
+  '/notifications': typeof NotificationsRouteWithChildren
   '/notifications/unsubscribe': typeof NotificationsUnsubscribeRoute
   '/proposal/$proposalId': typeof ProposalProposalIdRoute
+  '/notifications/': typeof NotificationsIndexRoute
   '/api/cron/notifications': typeof ApiCronNotificationsRoute
   '/api/notifications/telegram': typeof ApiNotificationsTelegramRoute
 }
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
   '/delegation': typeof DelegationRoute
-  '/notifications': typeof NotificationsRoute
   '/notifications/unsubscribe': typeof NotificationsUnsubscribeRoute
   '/proposal/$proposalId': typeof ProposalProposalIdRoute
+  '/notifications': typeof NotificationsIndexRoute
   '/api/cron/notifications': typeof ApiCronNotificationsRoute
   '/api/notifications/telegram': typeof ApiNotificationsTelegramRoute
 }
@@ -75,9 +84,10 @@ export interface FileRoutesById {
   __root__: typeof rootRouteImport
   '/': typeof IndexRoute
   '/delegation': typeof DelegationRoute
-  '/notifications': typeof NotificationsRoute
+  '/notifications': typeof NotificationsRouteWithChildren
   '/notifications/unsubscribe': typeof NotificationsUnsubscribeRoute
   '/proposal/$proposalId': typeof ProposalProposalIdRoute
+  '/notifications/': typeof NotificationsIndexRoute
   '/api/cron/notifications': typeof ApiCronNotificationsRoute
   '/api/notifications/telegram': typeof ApiNotificationsTelegramRoute
 }
@@ -89,15 +99,16 @@ export interface FileRouteTypes {
     | '/notifications'
     | '/notifications/unsubscribe'
     | '/proposal/$proposalId'
+    | '/notifications/'
     | '/api/cron/notifications'
     | '/api/notifications/telegram'
   fileRoutesByTo: FileRoutesByTo
   to:
     | '/'
     | '/delegation'
-    | '/notifications'
     | '/notifications/unsubscribe'
     | '/proposal/$proposalId'
+    | '/notifications'
     | '/api/cron/notifications'
     | '/api/notifications/telegram'
   id:
@@ -107,6 +118,7 @@ export interface FileRouteTypes {
     | '/notifications'
     | '/notifications/unsubscribe'
     | '/proposal/$proposalId'
+    | '/notifications/'
     | '/api/cron/notifications'
     | '/api/notifications/telegram'
   fileRoutesById: FileRoutesById
@@ -114,8 +126,7 @@ export interface FileRouteTypes {
 export interface RootRouteChildren {
   IndexRoute: typeof IndexRoute
   DelegationRoute: typeof DelegationRoute
-  NotificationsRoute: typeof NotificationsRoute
-  NotificationsUnsubscribeRoute: typeof NotificationsUnsubscribeRoute
+  NotificationsRoute: typeof NotificationsRouteWithChildren
   ProposalProposalIdRoute: typeof ProposalProposalIdRoute
   ApiCronNotificationsRoute: typeof ApiCronNotificationsRoute
   ApiNotificationsTelegramRoute: typeof ApiNotificationsTelegramRoute
@@ -144,12 +155,19 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof NotificationsRouteImport
       parentRoute: typeof rootRouteImport
     }
+    '/notifications/': {
+      id: '/notifications/'
+      path: '/'
+      fullPath: '/notifications/'
+      preLoaderRoute: typeof NotificationsIndexRouteImport
+      parentRoute: typeof NotificationsRoute
+    }
     '/notifications/unsubscribe': {
       id: '/notifications/unsubscribe'
-      path: '/notifications/unsubscribe'
+      path: '/unsubscribe'
       fullPath: '/notifications/unsubscribe'
       preLoaderRoute: typeof NotificationsUnsubscribeRouteImport
-      parentRoute: typeof rootRouteImport
+      parentRoute: typeof NotificationsRoute
     }
     '/proposal/$proposalId': {
       id: '/proposal/$proposalId'
@@ -175,11 +193,24 @@ declare module '@tanstack/react-router' {
   }
 }
 
+interface NotificationsRouteChildren {
+  NotificationsUnsubscribeRoute: typeof NotificationsUnsubscribeRoute
+  NotificationsIndexRoute: typeof NotificationsIndexRoute
+}
+
+const NotificationsRouteChildren: NotificationsRouteChildren = {
+  NotificationsUnsubscribeRoute: NotificationsUnsubscribeRoute,
+  NotificationsIndexRoute: NotificationsIndexRoute,
+}
+
+const NotificationsRouteWithChildren = NotificationsRoute._addFileChildren(
+  NotificationsRouteChildren,
+)
+
 const rootRouteChildren: RootRouteChildren = {
   IndexRoute: IndexRoute,
   DelegationRoute: DelegationRoute,
-  NotificationsRoute: NotificationsRoute,
-  NotificationsUnsubscribeRoute: NotificationsUnsubscribeRoute,
+  NotificationsRoute: NotificationsRouteWithChildren,
   ProposalProposalIdRoute: ProposalProposalIdRoute,
   ApiCronNotificationsRoute: ApiCronNotificationsRoute,
   ApiNotificationsTelegramRoute: ApiNotificationsTelegramRoute,
